+++ react-container/src/auth/__tests__/auth.test.tsx
@@ -1,3 +1,4 @@
+import { renderHook } from '@testing-library/react-hooks';
 import { getAuthHeaders } from '../auth';
 
 const mock = jest.fn().mockImplementation(() => {
@@ -8,10 +9,13 @@ jest.mock('utils/log', mock);
 
 describe('getAuthHeaders', () => {
   it('should be ok with nothing set', () => {
-    expect(getAuthHeaders()).toEqual({ 'api-key': '' });
+    const { result } = renderHook(() => getAuthHeaders());
+
+    expect(result.current).toEqual({ 'api-key': '' });
   });
   it('should set custom api header', () => {
     process.env.REACT_APP_API_KEY = 'test';
-    expect(getAuthHeaders('x-api-key')).toEqual({ 'x-api-key': 'test' });
+    const { result } = renderHook(() => getAuthHeaders());
+    expect(result.current).toEqual({ 'api-key': 'test' });
   });
 });