+++ react-container/src/utils/sidecar.ts
@@ -1,6 +1,6 @@
 type SidecarConfig = {
   __sidecarFormatVersion: number;
-  AADClientID?: string;
+  aadApplicationId?: string;
   AADTenantID?: string;
   applicationId: string;
   applicationName: string;
@@ -39,9 +39,7 @@ const DEFAULT_SIDECAR: SidecarConfig = {
   AADTenantID: 'reactdemoapp.onmicrosoft.com',
 
   // created with terraform:
-  AADClientID: '7e8b0cb9-c1bf-46df-bf79-13cb1b19a8f8', // Demo App: bluefield (staging)
-  // AADClientID: '52106ea0-e93b-40bc-99dc-5f7deacceda1', // Demo App: ew1 (staging)
-  // AADClientID: 'e0e9ac77-f9cc-49ca-9bb2-eef01313751b', // Demo App: greenfield (staging)
+  aadApplicationId: '7e8b0cb9-c1bf-46df-bf79-13cb1b19a8f8', // Demo App: bluefield (staging)
 };
 
 export const getSidecar = (): SidecarConfig => {